Fix some file not found exceptions occuring if ingestion is slow to clean up empty log files.

+++ sli/acceptance-tests/test/features/ingestion/features/step_definitions/ingestion_steps.rb
@@ -2763,14 +2763,16 @@ end
 
 After do
   if (!@landing_zone_path.nil?)
-          Dir.foreach(@landing_zone_path) do |entry|
-              if (entry.rindex("warn.") || entry.rindex("error."))
-                   STDOUT.puts "Error\/Warnings File detected = " + @landing_zone_path + entry
-                   STDOUT.puts "File contents follow:"
-                   STDOUT.puts File.open(@landing_zone_path + entry).read
-              end
-          end
+    Dir.foreach(@landing_zone_path) do |entry|
+      if (entry.rindex("warn.") || entry.rindex("error."))
+        if File.exists?(@landing_zone_path + entry)
+          STDOUT.puts "Error\/Warnings File detected = " + @landing_zone_path + entry
+          STDOUT.puts "File contents follow:"
+          STDOUT.puts File.open(@landing_zone_path + entry).read
+        end
       end
+    end
+  end
   cleanTenants()
   @conn.close if @conn != nil
   @batchConn.close if @batchConn != nil
